Changes:  - Grouping of the UI elements and the complex UI<->User interaction was separated from the `MainPanel` class and moved into a `UIHelper` class.
